renamed module tweakreg -> tweak
