+++ drizzlepac/__init__.py
@@ -62,7 +62,7 @@ from . import pixtopix
 # The following modules are for 'tweakreg' and are included here to make
 # it easier to get to this code interactively
 try:
-    from drizzlepac.tweakreg import tweakreg, imagefindpars, refimagefindpars, mapreg, catalogs, photeq, tweakutils
+    from drizzlepac.tweak import tweakreg, imagefindpars, refimagefindpars, mapreg, catalogs, photeq, tweakutils
     from . import imgclasses
 except ImportError as e :
     print('The libraries needed for "tweakreg" were not available!')
@@ -74,7 +74,7 @@ from . import updatenpol
 from . import buildwcs
 
 # This module supports applying WCS from _drz to _flt files
-from drizzlepac.tweakreg import tweakback
+from drizzlepac.tweak import tweakback
 
 # This module enables users to replace NaNs in images with another value easily
 from . import pixreplace
