+++ drizzlepac/tweak/findobj.py
