+++ drizzlepac/tweak/imagefindpars.py
