+++ drizzlepac/tweak/pars/imagefindpars.cfg
