+++ drizzlepac/tweak/pars/imagefindpars.cfgspc
