+++ drizzlepac/tweak/pars/mapreg.cfg
