+++ drizzlepac/tweak/pars/mapreg.cfgspc
