+++ drizzlepac/tweak/pars/photeq.cfg
