+++ drizzlepac/tweak/pars/photeq.cfgspc
