+++ drizzlepac/tweak/pars/tweakreg.cfg
