+++ drizzlepac/tweak/regfilter.py
